requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pyyaml>=6.0
anthropic>=0.40.0
httpx>=0.27.0
python-multipart>=0.0.12
